Various fixes - add constraints, add reclaim waiting period, fix casts (#63)
Address audit feedback:

- Remove `reclaim_event_account` instruction
- Since the unique nonce is no longer encoded into the message until
attestation time, we cannot reliably close a MessageSent event account
due to the missing uniqueness. This adds a 5 day waiting period that
callers must wait before closing the account. This means as long as the
off-chain attestation service (iris) is not > 5 days behind, it does not
have to worry about messages being closed before being read.
- Burn remaining balance of tokens in the custody token account in
`remove_local_token`
- This prevents us from being frontrun which could disallow removing the
local token account
- Add missing check to `set_token_controller`
- Other admin role setters check that the new token controller !=
existing, this adds that check
- Fix `get_min_fee_amount` fee calculation
  - Casts all values to `u128` during calculations to avoid u64 overflow
- Add check to `MessageTransmitter::send_message` that destination
domain != local domain to match EVM functionality
- Fixes data length of CPI call into handle_receive_un/finalized_message
- Adds constraint to force fee_recipient_token_account to be the
fee_recipient ATA rather than any token account owned by the
fee_recipient

+++ programs/v2/token-messenger-minter-v2/src/token_messenger_v2/state.rs
@@ -69,12 +69,13 @@ impl TokenMessenger {
 
         require_gt!(amount, 1, TokenMessengerError::InvalidAmount);
 
-        let fee = amount
-            .checked_mul(self.min_fee as u64)
+        // Cast amount and fees to 128 to avoid overflow, then cast result back to u64 when returning
+        let fee = (amount as u128)
+            .checked_mul(self.min_fee as u128)
             .ok_or(TokenMessengerError::MinFeeAmountOverflow)?
-            / MIN_FEE_MULTIPLIER;
+            / MIN_FEE_MULTIPLIER as u128;
 
-        Ok(fee.max(1))
+        Ok(fee.max(1) as u64)
     }
 }
